requirements.txt
fastapi
uvicorn
python-pptx
sqlalchemy
aiosqlite
python-multipart
jinja2
scikit-learn
joblib
pytest
pytest-cov
python-dotenv
